+++ src/Entity/WorkerGroup.php
@@ -1,0 +1,78 @@
+<?php
+
+namespace App\Entity;
+
+use App\Repository\WorkerGroupRepository;
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+use Doctrine\ORM\Mapping as ORM;
+
+#[ORM\Entity(repositoryClass: WorkerGroupRepository::class)]
+#[ORM\Table(name: 'worker_group')]
+class WorkerGroup
+{
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    private ?int $id = null;
+
+    #[ORM\Column(length: 255, unique: true)]
+    private ?string $name = null;
+
+    /**
+     * @var Collection<int, Worker>
+     */
+    #[ORM\ManyToMany(targetEntity: Worker::class, inversedBy: 'workerGroups')]
+    private Collection $workers;
+
+    public function __construct()
+    {
+        $this->workers = new ArrayCollection();
+    }
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getName(): ?string
+    {
+        return $this->name;
+    }
+
+    public function setName(string $name): static
+    {
+        $this->name = $name;
+
+        return $this;
+    }
+
+    public function __toString(): string
+    {
+        return (string) ($this->name ?? $this->id);
+    }
+
+    /**
+     * @return Collection<int, Worker>
+     */
+    public function getWorkers(): Collection
+    {
+        return $this->workers;
+    }
+
+    public function addWorker(Worker $worker): static
+    {
+        if (!$this->workers->contains($worker)) {
+            $this->workers->add($worker);
+        }
+
+        return $this;
+    }
+
+    public function removeWorker(Worker $worker): static
+    {
+        $this->workers->removeElement($worker);
+
+        return $this;
+    }
+}