+++ src/Model/Planning/Assignee.php
@@ -12,11 +12,13 @@ class Assignee
     public ArrayCollection $projects;
     /** @var ArrayCollection<string, SprintSum> */
     public ArrayCollection $sprintSums;
+    public readonly ?float $weekNorm;
 
-    public function __construct(string $key, string $displayName)
+    public function __construct(string $key, string $displayName, ?float $weekNorm = null)
     {
         $this->key = $key;
         $this->displayName = $displayName;
+        $this->weekNorm = $weekNorm;
         $this->projects = new ArrayCollection();
         $this->sprintSums = new ArrayCollection();
     }